backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
SQLAlchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
pydantic==2.9.2
pydantic-settings==2.6.1
python-multipart==0.0.17
boto3==1.35.71
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
